requirements.txt
streamlit
langchain
langchain-openai
langchain-community
faiss-cpu
pypdf
python-dotenv
pandas
numpy
openai
python-docx